test/web-platform-tests/to-upstream/css/css-color/nested-color-mix-with-currentcolor.html

<!doctype html>
<title>CSS Color: Nested color-mix() with currentColor</title>
<link rel="help" href="https://www.w3.org/TR/css-color-5/#color-mix">
<link rel="help" href="https://www.w3.org/TR/css-color-5/#resolving-color-values">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #parent {
    color: red;
    background-color: color-mix(in lch, color-mix(in lch, black, currentColor), black);
  }
  #parent2 {
    color: red;
    background-color: color-mix(in srgb, color-mix(in srgb, black, currentColor), black);
  }
  #child {
    color: black;
    background-color: inherit;
  }
  #child2 {
    color: white;
    background-color: inherit;
  }
</style>
<div id="parent">
  <div id="child"></div>
</div>
<div id="parent2">
  <div id="child2"></div>
</div>
<script>
  "use strict";
  const child = document.getElementById("child");
  const child2 = document.getElementById("child2");

  test(() => {
    assert_equals(getComputedStyle(child).backgroundColor, "lch(0 0 none)");
  }, "Nested color-mix function with inner currentColor should inherit unresolved");

  test(() => {
    assert_equals(getComputedStyle(child2).backgroundColor, "color(srgb 0.25 0.25 0.25)");
  }, "Nested color-mix function with inner currentColor should inherit unresolved; part2");
</script>
